Apply final readability fixes in innings and win calculations
Agent-Logs-Url: https://github.com/sooubh/GULLY-CRICKET/sessions/50b034f5-e6f0-4e18-8142-88bb24d6abe5

+++ lib/features/scoring/domain/engines/rule_engine.dart
@@ -55,8 +55,7 @@ class RuleEngine {
     int? target,
   }) {
     final totalBatters = battingPlayers.length;
-    final wicketsToEnd =
-        (rules.lastManBatsAlone ? totalBatters : totalBatters - 1).clamp(0, 11).toInt();
+    final wicketsToEnd = (rules.lastManBatsAlone ? totalBatters : totalBatters - 1).clamp(0, 11);
     final unavailableBatters = battingPlayers
         .where((player) => player.isOut || player.isRetiredHurt || (player.isRetired && !rules.reEntryAllowed))
         .length;
